djangorest/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-environ==0.4.5
psycopg2-binary==2.9.1